added session file
added the session.file to store every database/session functions
and updated them
