--- src/handlers/verifUser.go
@@ -1,1 +1,0 @@
-package handlers
